Updated texture view implementation
