+++ packages/hmssdk_flutter/lib/src/ui/meeting/hms_texture_view_controller.dart
@@ -9,21 +9,31 @@ import 'package:hmssdk_flutter/hmssdk_flutter.dart';
 import 'package:hmssdk_flutter/src/enum/hms_video_view_event.dart';
 import 'package:hmssdk_flutter/src/service/platform_service.dart';
 
-///[HMSVideoViewController] is used to control the video view. It helps in controlling addTrack, removeTrack functionalities manually.
+///[HMSTextureViewController] is used to control the video view. It helps in controlling addTrack, removeTrack functionalities manually.
 ///It is useful in custom usecases where you wish to control the addTrack and removeTrack functionalities on your own.
-class HMSVideoViewController {
+///Please note that if you control the view creation, addTrack etc. on application, then application has the responsibility
+///to release the texture view as well by calling [disposeTextureView]
+class HMSTextureViewController {
   ///[_textureId] is the unique id of the texture view
   int? _textureId;
 
   ///getter for [_textureId]
   int? get textureId => _textureId;
 
+  ///[_height] is the height of the view
   int? _height;
+
+  ///[_width] is the width of the view
   int? _width;
+
+  ///[_updateViewCallback] is the callback required for refreshing the view when certain
+  ///properties of the view changes.
   Function? _updateViewCallback;
+
+  ///[aspectRatio] is the aspect ratio of the view
   double aspectRatio = 1;
 
-  HMSVideoViewController(
+  HMSTextureViewController(
       {HMSVideoTrack? track,
       bool addTrackByDefault = true,
       bool? disableAutoSimulcastLayerSelect = false}) {
@@ -33,6 +43,10 @@ class HMSVideoViewController {
         disableAutoSimulcastLayerSelect: disableAutoSimulcastLayerSelect);
   }
 
+  ///[createTextureView] is used to create the texture view. It takes [track] as an optional parameter.
+  ///If [track] is provided, then it will add the track to the texture view by default.
+  ///If [addTrackByDefault] is set to true, then it will add the track to the texture view by default.
+  ///If [disableAutoSimulcastLayerSelect] is set to true, then it will disable the auto simulcast layer selection.
   void createTextureView(
       {HMSTrack? track,
       bool addTrackByDefault = true,
@@ -57,11 +71,16 @@ class HMSVideoViewController {
     }
   }
 
+  ///[setCallbackMethod] is used to set the callback method for the texture view.
+  ///This callback method is used to refresh the view when certain properties of the view changes.
   void setCallbackMethod(Function callback) {
     log("VKohli Calling setCallbackMethod -> $callback");
     _updateViewCallback = callback;
   }
 
+  ///[disposeTextureView] is used to dispose the texture view.
+  ///It is the responsibility of the application to dispose the texture view if the controller is created
+  ///by the application.
   void disposeTextureView() async {
     log("VKohli Calling disposeTextureView");
     var result = await PlatformService.invokeMethod(
@@ -72,6 +91,8 @@ class HMSVideoViewController {
     }
   }
 
+  ///[addTrack] is used to add the track to the texture view.
+  ///If [disableAutoSimulcastLayerSelect] is set to true, then it will disable the auto simulcast layer selection.
   void addTrack(
       {required HMSVideoTrack track,
       bool? disableAutoSimulcastLayerSelect}) async {
@@ -86,6 +107,8 @@ class HMSVideoViewController {
     });
   }
 
+  ///[_setDisplayResolution] is used to set the display resolution of the texture view.
+  ///It is used internally by the SDK.
   void _setDisplayResolution({required int height, required int width}) {
     PlatformService.invokeMethod(PlatformMethod.setDisplayResolution,
         arguments: {
@@ -95,11 +118,13 @@ class HMSVideoViewController {
         });
   }
 
+  ///[removeTrack] is used to remove the track from the texture view.
   void removeTrack() {
     PlatformService.invokeMethod(PlatformMethod.removeTrack,
         arguments: {"texture_id": textureId.toString()});
   }
 
+  ///[setHeightWidth] is used to set the height and width of the texture view.
   void setHeightWidth({required double height, required double width}) {
     if (_height != height.toInt() || _width != width.toInt()) {
       log("VKohli Calling setHeightWidth-> height: $height, width: $width");
@@ -109,6 +134,8 @@ class HMSVideoViewController {
     }
   }
 
+  ///[_eventListener] is the callback method for the texture view.
+  ///We get the native callbacks like [onResolutionChanged] from the texture view.
   void _eventListener(dynamic event) {
     log("VKohli HMSVideoView Event Fired $event");
