+++ packages/hmssdk_flutter/lib/src/ui/meeting/hms_video_view.dart
@@ -1,9 +1,9 @@
 // Dart imports:
 import 'dart:io' show Platform;
-import 'dart:math';
 
 // Flutter imports:
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart' show StandardMessageCodec;
 
 // Project imports:
@@ -34,11 +34,6 @@ import 'package:hmssdk_flutter/hmssdk_flutter.dart';
 /// **key** - [key] property can be used to forcefully rebuild the video widget by setting a unique key everytime.
 /// Similarly to avoid rebuilding the key should be kept the same for particular HMSVideoView.
 ///
-/// **addTrackByDefault** - To call addTrack by default as HMSVideoView is attached to the tree. Default value is [true]
-///
-/// **controller** - To control the video view, this is useful for custom usecases when you wish to control the addTrack and removeTrack
-/// track functionalities on your own.
-///
 /// Refer [HMSVideoView guide here](https://www.100ms.live/docs/flutter/v2/features/render-video)
 class HMSVideoView extends StatelessWidget {
   /// This will render video with trackId present in the track
@@ -58,13 +53,6 @@ class HMSVideoView extends StatelessWidget {
 
   /// [disableAutoSimulcastLayerSelect] - To disable auto simulcast (Adaptive Bitrate)
   final bool disableAutoSimulcastLayerSelect;
-
-  /// [addTrackByDefault] - To call addTrack by default as HMSVideoView is attached to the tree. Default value is [true]
-  final bool addTrackByDefault;
-
-  /// [controller] - To control the video view, this is useful for custom usecases when you wish to control the addTrack and removeTrack
-  /// track functionalities on your own.
-  final HMSVideoViewController? controller;
 
   ///100ms HMSVideoView
   ///
@@ -100,9 +88,7 @@ class HMSVideoView extends StatelessWidget {
           "matchParent is not longer necessary and will be removed in future version")
       this.matchParent = true,
       this.scaleType = ScaleType.SCALE_ASPECT_FIT,
-      this.disableAutoSimulcastLayerSelect = false,
-      this.addTrackByDefault = true,
-      this.controller})
+      this.disableAutoSimulcastLayerSelect = false})
       : super(key: key);
 
   @override
@@ -113,20 +99,17 @@ class HMSVideoView extends StatelessWidget {
       setMirror: setMirror,
       scaleType: this.scaleType,
       disableAutoSimulcastLayerSelect: disableAutoSimulcastLayerSelect,
-      addTrackByDefault: addTrackByDefault,
-      controller: controller,
     );
   }
 }
 
-class _PlatformView extends StatefulWidget {
+class _PlatformView extends StatelessWidget {
   final HMSTrack track;
+
   final bool setMirror;
   final bool matchParent;
   final ScaleType scaleType;
   final bool disableAutoSimulcastLayerSelect;
-  final bool addTrackByDefault;
-  final HMSVideoViewController? controller;
 
   _PlatformView(
       {Key? key,
@@ -134,91 +117,42 @@ class _PlatformView extends StatefulWidget {
       this.setMirror = false,
       this.matchParent = true,
       required this.scaleType,
-      this.disableAutoSimulcastLayerSelect = false,
-      this.addTrackByDefault = true,
-      this.controller})
+      this.disableAutoSimulcastLayerSelect = false})
       : super(key: key);
 
-  @override
-  State<_PlatformView> createState() => _PlatformViewState();
-}
-
-class _PlatformViewState extends State<_PlatformView> {
-  HMSVideoViewController? viewController;
-
   void onPlatformViewCreated(int id) {}
-
-  @override
-  void initState() {
-    if (widget.controller == null) {
-      viewController =
-          HMSVideoViewController(track: widget.track as HMSVideoTrack);
-    } else {
-      viewController = widget.controller;
-    }
-    viewController?.setCallbackMethod(setView);
-    super.initState();
-  }
-
-  void setView() {
-    if (mounted) {
-      setState(() {});
-    }
-  }
-
-  @override
-  void dispose() {
-    if (widget.controller == null) {
-      viewController?.disposeTextureView();
-    }
-    super.dispose();
-  }
 
   @override
   Widget build(BuildContext context) {
     ///AndroidView for android it uses surfaceRenderer provided internally by webrtc.
     if (Platform.isAndroid) {
-      return viewController?.textureId == null
-          ? SizedBox()
-          : LayoutBuilder(
-              builder: (context, constraints) {
-                viewController?.setHeightWidth(
-                    height: constraints.maxHeight, width: constraints.maxWidth);
-                return Center(
-                  child: widget.scaleType != ScaleType.SCALE_ASPECT_FIT
-                      ? Container(
-                          width: constraints.maxWidth,
-                          height: constraints.maxHeight,
-                          child: HMSTextureView(
-                            scaleType: widget.scaleType,
-                            viewController: viewController,
-                            constraints: constraints,
-                            setMirror: widget.setMirror,
-                          ),
-                        )
-                      : HMSTextureView(
-                          scaleType: widget.scaleType,
-                          viewController: viewController,
-                          constraints: constraints,
-                          setMirror: widget.setMirror,
-                        ),
-                );
-              },
-            );
+      return AndroidView(
+        hitTestBehavior: PlatformViewHitTestBehavior.transparent,
+        viewType: 'HMSVideoView',
+        onPlatformViewCreated: onPlatformViewCreated,
+        creationParamsCodec: StandardMessageCodec(),
+        creationParams: {
+          'track_id': track.trackId,
+          'set_mirror': track.source != "REGULAR" ? false : setMirror,
+          'scale_type': scaleType.value,
+          'match_parent': matchParent,
+          'disable_auto_simulcast_layer_select': disableAutoSimulcastLayerSelect
+        },
+        gestureRecognizers: {},
+      );
     } else if (Platform.isIOS) {
       ///UIKitView for ios it uses VideoView provided by 100ms ios_sdk internally.
       return UiKitView(
+        hitTestBehavior: PlatformViewHitTestBehavior.transparent,
         viewType: 'HMSFlutterPlatformView',
         onPlatformViewCreated: onPlatformViewCreated,
         creationParamsCodec: StandardMessageCodec(),
         creationParams: {
-          'track_id': widget.track.trackId,
-          'set_mirror':
-              widget.track.source != "REGULAR" ? false : widget.setMirror,
-          'scale_type': widget.scaleType.value,
-          'match_parent': widget.matchParent,
-          'disable_auto_simulcast_layer_select':
-              widget.disableAutoSimulcastLayerSelect
+          'track_id': track.trackId,
+          'set_mirror': track.source != "REGULAR" ? false : setMirror,
+          'scale_type': scaleType.value,
+          'match_parent': matchParent,
+          'disable_auto_simulcast_layer_select': disableAutoSimulcastLayerSelect
         },
         gestureRecognizers: {},
       );
@@ -226,41 +160,5 @@ class _PlatformViewState extends State<_PlatformView> {
       throw UnimplementedError(
           'Video View is not implemented for this platform ${Platform.localHostname}');
     }
-  }
-}
-
-class HMSTextureView extends StatelessWidget {
-  const HMSTextureView(
-      {Key? key,
-      required this.scaleType,
-      required this.viewController,
-      required this.constraints,
-      required this.setMirror})
-      : super(key: key);
-
-  final ScaleType scaleType;
-  final HMSVideoViewController? viewController;
-  final BoxConstraints constraints;
-  final bool setMirror;
-
-  @override
-  Widget build(BuildContext context) {
-    return FittedBox(
-      clipBehavior: Clip.hardEdge,
-      fit: scaleType == ScaleType.SCALE_ASPECT_FIT
-          ? BoxFit.contain
-          : BoxFit.cover,
-      child: SizedBox(
-        width: (constraints.maxHeight *
-            ((viewController != null) ? viewController!.aspectRatio : 1)),
-        height: constraints.maxHeight,
-        child: Center(
-          child: Transform(
-              transform: Matrix4.identity()..rotateY(setMirror ? -pi : 0.0),
-              alignment: FractionalOffset.center,
-              child: Texture(textureId: viewController!.textureId!)),
-        ),
-      ),
-    );
   }
 }
